requirements.txt

pandas
folium
polyline
Shapely
pytz

gspread
oauth2client
gspread-dataframe
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
googlemaps


mygeotab

boto3


psycopg2-binary


flask
<<<<<<< HEAD
gunicorn
=======
botocore
>>>>>>> ebb7e5a (updates MVP DVI)
